simple_requirements.txt
flask==3.0.0
flask-cors==4.0.0
pyjwt==2.8.0
gunicorn==21.2.0 